Add extra BaseRestClient without a fixed ApiService to it.
